+++ src/model/playlist.rs
@@ -61,17 +61,6 @@ pub(crate) enum PlaylistItemType {
     SeriesInfo = 4,
 }
 
-impl PlaylistItemType {
-    pub(crate) fn is_same(&self, value: &str) -> bool {
-        match self {
-            PlaylistItemType::Live => value.eq_ignore_ascii_case("live"),
-            PlaylistItemType::Movie => value.eq_ignore_ascii_case("movie"),
-            PlaylistItemType::Series => value.eq_ignore_ascii_case("series"),
-            PlaylistItemType::SeriesInfo => value.eq_ignore_ascii_case("series-info"),
-        }
-    }
-}
-
 impl Display for PlaylistItemType {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", match self {