+++ src/utils/config_reader.rs
@@ -20,8 +20,10 @@ pub(crate) fn read_mappings(args_mapping: Option<String>, cfg: &mut Config) -> R
     match read_mapping(mappings_file.as_str()) {
         Ok(mappings) => {
             info!("Mappings File: {}", &mappings_file);
-            if mappings.is_none() { debug!("no mapping loaded"); }
-            cfg.set_mappings(mappings);
+            match mappings {
+                None => {debug!("no mapping loaded");}
+                Some(mappings_cfg) => {cfg.set_mappings(mappings_cfg);}
+            }
             Ok(())
         }
         Err(err) => Err(err),